[#1402] Move React embed test into its own package; scope legacy embed test
CI fixed: the React embed smoke test was failing because the file lived in
package apiserver_test, so `go test ./apiserver/...` also compiled
apiserver_with_frontend.go which has a //go:embed all:dist on the legacy
frontend bundle that the React workflow does not build.

- Move TestFrontendReactEmbed_* from apiserver_test to its own
  ./apiserver/frontendreactembed/ package (with a non-tagged doc.go so the
  package always exists for `go list ./...`).
- Update frontend-react-embed-smoke-test.yml to test
  ./apiserver/frontendreactembed/... — only frontend-react/dist is needed.
- Update frontend-embed-smoke-test.yml to test ./apiserver (no /...) so the
  new sub-package is not pulled in — only frontend/dist is needed.

Verified locally:
- go test -tags with_frontend ./apiserver/frontendreactembed — pass
- go test -tags with_frontend ./apiserver -run TestFrontendEmbed — pass
- go test ./apiserver/frontendreactembed/... (no tag) — no test files (ok)
- go mod tidy keeps require/replace lines for both bundles
